--- app/core/Routing.js
@@ -1,8 +1,0 @@
-export default (app, routes) => {
-    routes.forEach(item =>
-        app.use(item.path, item.middlewares, item.component)
-    );
-
-    return app;
-}
-
