--- app/index.js
@@ -1,3 +1,0 @@
-import app from './app.js';
-
-export default app;
