--- app/location/location.fixtures.js
@@ -1,202 +1,0 @@
-export var supportedCities = [
-    {
-        timezone: 1,
-        name: "europe/dublin",
-        city: "Dublin",
-        country: "Ireland",
-        drink: "Guinness",
-        note: "A dark Irish dry stout that originated in the brewery of Arthur Guinness at St. James's Gate, Dublin, Ireland, in 1759."
-    },
-    {
-        timezone: 2,
-        name: "europe/rome",
-        city: "Rome",
-        country: "Italia",
-        drink: "Gin Negroni",
-        note: "A cocktail made of one part gin, one part vermouth rosso, and one part Campari, garnished with orange peel.",
-    },
-    {
-        timezone: 3,
-        name: "europe/moscow",
-        city: "Moscow",
-        country: "Russia",
-        drink: "Russia",
-        note: "A clear distilled alcoholic beverage with different varieties originating in Poland and Russia.",
-    },
-    {
-        timezone: 4,
-        name: "asia/dubai",
-        city: "Dubai",
-        country: "United Arab Emirates",
-        drink: "Arabic Coffee (Gahwa)",
-        note: "A warm beverage made from roasting the coffee beans and blended with cardamom.",
-    },
-    {
-        timezone: 5,
-        name: "indian/maldives",
-        city: "Malé",
-        country: "Maldives",
-        drink: "The Maldivian Lady",
-        note: "A cocktail made from rum, apricot brandy, orange juice, pineapple juice, and splash of Grenadine syrup."
-    },
-    {
-        timezone: 6,
-        name: "asia/bishkek",
-        city: "Bishkek",
-        country: "Kyrgyzstan",
-        drink: "Jarma",
-        note: "A cold beverage made from fermented barley similar to beer.",
-    },
-    {
-        timezone: 7,
-        name: "asia/jakarta",
-        city: "Jakarta",
-        country: "Indonesia",
-        drink: "Bir Pletok",
-        note: "A refreshing drink made from a mixture of several spices, namely ginger, fragrant pandan leaves, and lemongrass.",
-    },
-    {
-        timezone: 8,
-        name: "asia/hong_kong",
-        city: "Hong Kong",
-        country: "Special Administrative Region",
-        drink: "Milk Tea",
-        note: "A drink served hot or cold made of black tea, evaporated milk, and sugar."
-    },
-    {
-        timezone: 9,
-        name: "asia/seoul",
-        city: "Seoul",
-        country: "Korea",
-        drink: "Soju",
-        note: "A clear, low-alcohol, distilled spirit made from rice, this beverage of Korean origin."
-    },
-    {
-        timezone: 9.5,
-        name: "australia/sydney",
-        city: "Sydney",
-        country: "Australia",
-        drink: "Lemon, lime and bitters (LLB)",
-        note: "A mixed drink made with lemonade, lime cordial, and Angostura bitters."
-    },
-    {
-        timezone: 11,
-        name: "asia/magadan",
-        city: "Magadan",
-        country: "Russia",
-        drink: "Vodka",
-        note: "A clear distilled alcoholic beverage with different varieties originating in Poland and Russia."
-    },
-    {
-        timezone: 0,
-        name: "africa/casablanca",
-        city: "Casa Blanca",
-        country: "Morocco",
-        drink: "Moroccan mint tea",
-        note: "A green tea prepared with spearmint leaves and sugar.",
-    },
-    {
-        timezone: 13,
-        name: "pacific/samoa",
-        city: "Samoa",
-        country: "Independent State of Samoa",
-        drink: "Ginger Mojito",
-        note: "A cocktail made of rum, sugar, lime juice, soda water, mint, and ginger."
-    },
-    {
-        timezone: 14,
-        name: "pacific/honolulu",
-        city: "Honolulu",
-        country: "United States of America",
-        drink: "Mai Tai",
-        note: "A cocktail made with rum, Curaçao liqueur, orgeat syrup, and lime juice."
-    },
-    {
-        timezone: -9,
-        name: "pacific/gambier",
-        city: "Gambier",
-        country: "South Australia",
-        drink: "Wine",
-        note: "The Mount Gambier wine region specialties include Pinot Noir, Merlot, and Chardonnay."
-    },
-    {
-        timezone: -8,
-        name: "america/yakutat",
-        city: "Yakutat",
-        country: "United States of America",
-        drink: "Beer",
-        note: "Anchorage is filled with Breweries."
-    },
-    {
-        timezone: -7,
-        name: "america/los_angeles",
-        city: "Los Angeles",
-        country: "United States of America",
-        drink: "Mimosa",
-        note: "A brunch cocktail is made with champagne and chilled orange juice."
-    },
-    {
-        timezone: -6,
-        name: "america/mazatlan",
-        city: "Mazatlan",
-        country: "Mexico",
-        drink: "Horchata",
-        note: "A cold beverage made of rice milk, cinnamon, and vanilla."
-    },
-    {
-        timezone: -5,
-        name: "america/lima",
-        city: "Lima",
-        country: "Peru",
-        drink: "Pisco Sour",
-        note: "A cocktail made from Pisco liquor, freshly squeezed lime juice, simple syrup, ice, egg white, and Angostura bitters."
-    },
-    {
-        timezone: -4,
-        name: "america/new_york",
-        city: "New York",
-        country: "United States of America",
-        drink: "Manhattan",
-        note: "A cocktail made with whiskey, sweet vermouth, and bitters."
-    },
-    {
-        timezone: -3,
-        name: "america/argentina",
-        city: "Buenos Aires", //mismatching city
-        country: "Argentina",
-        drink: "Mate",
-        note: "A tea made from the soaking dried yerba mate leaves in hot water."
-    },
-    {
-        timezone: -2,
-        name: "atlantic/south_georgia",
-        city: "Fernando de Noronha", //mismatching city
-        country: "Brazil",
-        drink: "Caipirinha",
-        note: "Is Brazil's national cocktail, made with cachaça (sugarcane hard liquor), sugar, and lime."
-    },
-    {
-        timezone: -2,
-        name: "america/st_johns",
-        city: "St Johns",
-        country: "United States of America",
-        drink: "Painkiller",
-        note: "A rum cocktail made with  pineapple juice, cream of coconut and orange juice, with a generous amount of fresh nutmeg on top."
-    },
-    {
-        timezone: -1,
-        name: "atlantic/cape_verde",
-        city: "Praia",
-        country: "Cape Verde",
-        drink: "Grogue",
-        note: "The national drink of Cape Verde, It is rum distilled from sugar cane."
-    },
-    {
-        timezone: 12,
-        name: "pacific/fiji",
-        city: "Suva",
-        country: "Fiji",
-        drink: "Kava",
-        note: "A staple of the Fijian culture, a drink made from ground-up Yaqona root mixed with water."
-    }
-]
