--- app/main/main.routes.js
@@ -1,7 +1,0 @@
-import Foo from './Foo.js';
-
-const routes = [
-    { path: '/', component: Foo, middleware: []}
-];
-
-export default routes;
