--- app/user/passport-setup.js
@@ -1,36 +1,0 @@
-import passport from './passport';
-import { Strategy as GoogleStrategy } from 'passport-google-oauth20';
-
-
-passport.serializeUser(function(user, done) {
-    /*
-    From the user take just the id (to minimize the cookie size) and just pass the id of the user
-    to the done callback
-    PS: You dont have to do it like this its just usually done like this
-    */
-    done(null, user);
-  });
-  
-passport.deserializeUser(function(user, done) {
-    /*
-    Instead of user this function usually recives the id 
-    then you use the id to select the user from the db and pass the user obj to the done callback
-    PS: You can later access this data in any routes in: req.user
-    */
-    done(null, user);
-});
-
-passport.use(new GoogleStrategy({
-    clientID: "265948777768-3u6d1h9kd31v0a2lpdbkaudpilrq6hgq.apps.googleusercontent.com",
-    clientSecret: "zCCOhuJJhqDgr6KnJNoXlTd8",
-    callbackURL: "http://localhost:3000/google/callback"
-  },
-  function(accessToken, refreshToken, profile, done) {
-    /*
-     use the profile info (mainly profile id) to check if the user is registerd in ur db
-     If yes select the user and pass him to the done callback
-     If not create the user and then select him and pass to callback
-    */
-    return done(null, profile);
-  }
-));
