+++ app/user/user.helpers.js
@@ -1,8 +1,6 @@
-import {G_OAUTH_CALLBACK_URL} from '../utilities/constants.js'
+export const adminLocationPage = (req, res) => res.redirect("/admin/location");;
 
-export const googleCallback = (req, res) => res.redirect(G_OAUTH_CALLBACK_URL);
-
-export const userInfo = (req, res) => res.send(req.user_json);
+export const userInfo = (req, res) => res.send(req.user._json);
 
 export const authFailure = (req, res) => res.sendStatus(401);
 
@@ -17,11 +15,11 @@ export const isLoggedIn = (req, res, next) => {
         "carlos.mendoza.91@gmail.com",
         "theweekendisneverover@gmail.com"
     ]
-    console.log(req.user)
-    if(req.user){
+    if(req.user) {
+        console.log(req.user)
         const userEmail = req.user._json.email;
         if(emails.includes(userEmail))
             next();
-    }
-    res.sendStatus(401);
+     } else
+        res.sendStatus(401);
 }
