+++ app/user/user.js
@@ -1,8 +1,8 @@
 import { URL_BASE } from '../utilities/constants.js';
-import router from './user.router.js'
+import userRouter from './user.router.js'
 
 const userRoutes = (app) => {
-    app.use(`${URL_BASE}/user`, router);
+    app.use(`${URL_BASE}/user`, userRouter);
     return app;
 }
