+++ app/utilities/pipe.js
@@ -1,0 +1,4 @@
+const _pipe = (a, b) => (arg) => b(a(arg));
+const pipe = (...ops) => ops.reduce(_pipe)
+
+export default pipe;
