--- app2/app.js
@@ -1,5 +1,0 @@
-import main from './main/index.js'
-
-const app = () => main();
-
-export default app;
