--- app2/main/index.js
@@ -1,3 +1,0 @@
-import main from './main.js'
-
-export default main;
