--- app2/user/index.js
@@ -1,3 +1,0 @@
-import userRoutes from './user.js';
-
-export default userRoutes;
